bug fixes: replace datastorehook to storeData and rename the file and placed it into utils folder
